OJ-3192 - Final re-organise
