+++ lambdas/common/package.json
@@ -10,7 +10,10 @@
     "deploy": "../../deploy.sh",
     "compile": "tsc"
   },
-  "dependencies": {},
+  "dependencies": {
+    "@aws-sdk/client-dynamodb": "^3.682.0",
+    "@aws-sdk/util-dynamodb": "3.788.0"
+  },
   "devDependencies": {
     "@types/jest": "^29.5.5",
     "@typescript-eslint/eslint-plugin": "^8.13.0",
