+++ lambdas/common/src/database/exceptions/errors.ts
