+++ lambdas/common/src/database/util/is-record-expired.ts
