+++ lambdas/common/tests/retry.test.ts
@@ -1,4 +1,4 @@
-import { withRetry } from "../src/retry";
+import { withRetry } from "../src/util/retry";
 
 jest.setTimeout(30000);
