--- lambdas/database/jest.config.ts
@@ -1,7 +1,0 @@
-import type { Config } from "jest";
-import baseConfig from "../../jest.config.base";
-
-export default {
-  ...baseConfig,
-  displayName: "lambdas/common",
-} satisfies Config;
