--- lambdas/database/package.json
@@ -1,30 +1,0 @@
-{
-  "name": "database",
-  "description": "",
-  "scripts": {
-    "lint": "eslint .",
-    "lint:fix": "eslint . --fix",
-    "unit": "jest --silent",
-    "test": "npm run unit --",
-    "test:coverage": "npm run unit -- --coverage",
-    "deploy": "../../deploy.sh",
-    "compile": "tsc"
-  },
-  "dependencies": {
-    "@aws-sdk/client-dynamodb": "^3.682.0",
-    "@aws-sdk/util-dynamodb": "3.788.0"
-  },
-  "devDependencies": {
-    "@types/jest": "^29.5.5",
-    "@typescript-eslint/eslint-plugin": "^8.13.0",
-    "@typescript-eslint/parser": "^8.13.0",
-    "eslint": "^8.53.0",
-    "eslint-config-prettier": "^9.1.0",
-    "eslint-plugin-prettier": "^5.2.1",
-    "jest": "^29.7.0",
-    "prettier": "^3.1.0",
-    "ts-jest": "^29.1.1",
-    "ts-node": "^10.9.2",
-    "typescript": "^5.3.2"
-  }
-}
